+++ lib/ble5stack/central/hci_tl.h
@@ -98,6 +98,7 @@
 
 // Low energy commands
 #define HCI_LE_READ_ADVERTISING_CHANNEL_TX_POWER          0x2007	//!< opcode of @ref HCI_LE_readAdvertisingChannelTxPower
+#define HCI_LE_SET_ADVERTISING_DATA                       0x2008
 
 /* HCI Status return types  */
 typedef enum